Handle newer client secrets stored without base64 encoding.
Webtasks using clients created after 12/06/2016 should define an additional
secret, AUTH0_SECRET_ENCODING=<encoding>, in addition to the AUTH0_CLIENT_ID,
AUTH0_CLIENT_SECRET, and AUTH0_DOMAIN secrets when the webtask is created.

If the AUTH0_SECRET_ENCODING secret is not provided, the <encoding> defaults
to 'base64' to retain backwards compatibility with clients created prior to
12/06/2016. Webtasks defined for clients created after 12/06/2016 should use
an <encoding> of 'utf8'.